Reform account edit page Fixes #86
