+++ lib/routes.dart
@@ -1,5 +1,6 @@
 import 'package:flow/entity/transaction.dart';
 import 'package:flow/l10n/extensions.dart';
+import 'package:flow/routes/account/account_edit_page.dart';
 import 'package:flow/routes/account_page.dart';
 import 'package:flow/routes/categories_page.dart';
 import 'package:flow/routes/category/category_edit_page.dart';
@@ -79,14 +80,23 @@ final router = GoRouter(
     ),
     GoRoute(
       path: '/account/new',
-      builder: (context, state) => const AccountPage.create(),
+      builder: (context, state) => const AccountEditPage.create(),
     ),
     GoRoute(
         path: '/account/:id',
-        builder: (context, state) => AccountPage.edit(
+        builder: (context, state) => AccountPage(
               accountId: int.tryParse(state.pathParameters["id"]!) ?? -1,
+              initialRange: TimeRange.tryParse(
+                state.uri.queryParameters["range"] ?? "",
+              ),
             ),
         routes: [
+          GoRoute(
+            path: 'edit',
+            builder: (context, state) => AccountEditPage(
+              accountId: int.tryParse(state.pathParameters["id"]!) ?? -1,
+            ),
+          ),
           GoRoute(
             path: 'transactions',
             builder: (context, state) => TransactionsPage.account(
